// libs/carry-app-services/src/enums/carry.enum.ts
export enum CARRY_EVENT_TYPES {
  CARRIER_ASSIGNED = 'CARRIER_ASSIGNED',
  CARRIER_UNASSIGNED = 'CARRIER_UNASSIGNED',
  CARRIER_MARKED_AS_PICKED_UP = 'CARRIER_MARKED_AS_PICKED_UP',
  CARRIER_MARKED_AS_DROPPED_OFF = 'CARRIER_MARKED_AS_DROPPED_OFF',
  DELIVERY_NEW_TRACKING_UPDATE = 'DELIVERY_NEW_TRACKING_UPDATE',
  DELIVERY_NEW_MESSAGE = 'DELIVERY_NEW_MESSAGE',
  DELIVERY_ADD_COMPLETED_PICTURE = 'DELIVERY_ADD_COMPLETED_PICTURE',

  DELIVERY_FROM_PERSON_ID_PICTURE_ADDED = 'DELIVERY_FROM_PERSON_ID_PICTURE_ADDED',
  DELIVERY_FROM_PERSON_SIG_PICTURE_ADDED = 'DELIVERY_FROM_PERSON_SIG_PICTURE_ADDED',
  DELIVERY_TO_PERSON_ID_PICTURE_ADDED = 'DELIVERY_TO_PERSON_ID_PICTURE_ADDED',
  DELIVERY_TO_PERSON_SIG_PICTURE_ADDED = 'DELIVERY_TO_PERSON_SIG_PICTURE_ADDED',

  DELIVERY_COMPLETED = 'DELIVERY_COMPLETED',
  DELIVERY_RETURNED = 'DELIVERY_RETURNED',

  CARRIER_LOCATION_REQUESTED = 'CARRIER_LOCATION_REQUESTED',
  CARRIER_LOCATION_REQUEST_CANCELED = 'CARRIER_LOCATION_REQUEST_CANCELED',
  CARRIER_LOCATION_REQUEST_ACCEPTED = 'CARRIER_LOCATION_REQUEST_ACCEPTED',
  CARRIER_LOCATION_REQUEST_DECLINED = 'CARRIER_LOCATION_REQUEST_DECLINED',
  CARRIER_LOCATION_SHARED = 'CARRIER_LOCATION_SHARED',
  CARRIER_LOCATION_UNSHARED = 'CARRIER_LOCATION_UNSHARED',
  CARRIER_LOCATION_UPDATED = 'CARRIER_LOCATION_UPDATED',

  TO_DELIVERY = 'TO_DELIVERY',

  NEW_DELIVERY_OWNER_REVIEW = 'NEW_DELIVERY_OWNER_REVIEW',
  NEW_DELIVERY_CARRIER_REVIEW = 'NEW_DELIVERY_CARRIER_REVIEW',

  NEW_DELIVERY_DISPUTE = 'NEW_DELIVERY_DISPUTE',
  NEW_DELIVERY_DISPUTE_LOG = 'NEW_DELIVERY_DISPUTE_LOG',
  DELIVERY_DISPUTE_SETTLEMENT_OFFER = 'NEW_DELIVERY_DISPUTE_SETTLEMENT_OFFER',
  DELIVERY_DISPUTE_SETTLEMENT_OFFER_CANCELED = 'DELIVERY_DISPUTE_SETTLEMENT_OFFER_CANCELED',
  DELIVERY_DISPUTE_SETTLEMENT_OFFER_ACCEPTED = 'DELIVERY_DISPUTE_SETTLEMENT_OFFER_ACCEPTED',
  DELIVERY_DISPUTE_SETTLEMENT_OFFER_DECLINED = 'DELIVERY_DISPUTE_SETTLEMENT_OFFER_DECLINED',
  NEW_DISPUTE_CUSTOMER_SERVICE_AGENT = 'NEW_DISPUTE_CUSTOMER_SERVICE_AGENT',
  STRIPE_ACCOUNT_VERIFIED = "STRIPE_ACCOUNT_VERIFIED",
  CARRIER_PAYOUT_CLAIMED = "CARRIER_PAYOUT_CLAIMED",
  CARRIER_DELIVERY_REQUEST = "CARRIER_DELIVERY_REQUEST",
  CARRIER_DELIVERY_REQUEST_CANCELED = "CARRIER_DELIVERY_REQUEST_CANCELED",
  CARRIER_DELIVERY_REQUEST_DECLINED = "CARRIER_DELIVERY_REQUEST_DECLINED",
  CARRIER_DELIVERY_REQUEST_ACCEPTED = "CARRIER_DELIVERY_REQUEST_ACCEPTED",
  CARRIER_REMOVED = "CARRIER_REMOVED",
  STRIPE_IDENTITY_VERIFIED = "STRIPE_IDENTITY_VERIFIED",
  STRIPE_IDENTITY_VERIFICATION_SESSION_CANCELED = "STRIPE_IDENTITY_VERIFICATION_SESSION_CANCELED",
  CARRIER_APPROACHING_PICKUP_LOCATION = "CARRIER_APPROACHING_PICKUP_LOCATION",
  CARRIER_APPROACHING_DROPOFF_LOCATION = "CARRIER_APPROACHING_DROPOFF_LOCATION"
}